Refactor DeploymentModeModal to use a base HeliumBottomSheet component
